Split unit tests to individual files
