+++ tests/BarcodeSvgTest.php
@@ -1,0 +1,14 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+
+class BarcodeSvgTest extends TestCase
+{
+    public function test_svg_barcode_generator_can_generate_ean_13_barcode()
+    {
+        $generator = new Picqer\Barcode\BarcodeGeneratorSVG();
+        $generated = $generator->getBarcode('081231723897', $generator::TYPE_EAN_13);
+
+        $this->assertStringEqualsFile('tests/verified-files/081231723897-ean13.svg', $generated);
+    }
+}
